+++ raven/__init__.py
@@ -1,1 +1,3 @@
+"""WPS processes related to GIS and hydrological modeling, namely the Raven hydrological modeling framework."""
+
 from .__version__ import __author__, __email__, __version__  # noqa: F401