+++ raven/processes/wps_generic_shape_properties.py
@@ -4,11 +4,12 @@ import logging
 import fiona
 from pyproj.crs import CRS
 from pywps import FORMATS, ComplexInput, ComplexOutput, LiteralInput, Process
-from ravenpy.utilities.analysis import geom_prop
-from ravenpy.utilities.checks import multipolygon_check, single_file_check
-from ravenpy.utilities.geo import geom_transform
-from ravenpy.utilities.io import archive_sniffer, crs_sniffer
 from shapely.geometry import shape
+
+from raven.utilities.analysis import geom_prop
+from raven.utilities.checks import multipolygon_check, single_file_check
+from raven.utilities.geo import geom_transform
+from raven.utilities.io import archive_sniffer, crs_sniffer
 
 LOGGER = logging.getLogger("PYWPS")