+++ raven/utilities/__init__.py
@@ -1,0 +1,12 @@
+# TODO: Merge these into one message.
+gis_import_error_message = (
+    "`{}` requires installation of the RavenPy GIS libraries. These can be installed using the"
+    " `pip install ravenpy[gis]` recipe or via Anaconda (`conda env update -n ravenpy-env -f environment.yml`)"
+    " from the RavenPy repository source files."
+)
+
+dev_import_error_message = (
+    "`{}` requires installation of the RavenPy development libraries. These can be installed using the"
+    " `pip install ravenpy[dev]` recipe or via Anaconda (`conda env update -n ravenpy-env -f environment.yml`)"
+    " from the RavenPy repository source files."
+)