+++ raven/utilities/geoserver.py
@@ -316,7 +316,7 @@ def get_raster_wcs(
             identifier=[layer],
             format="image/tiff",
             subsets=[(x, left, right), (y, down, up)],
-            timeout=120,
+            timeout=300,
         )
 
     except Exception: